Refactor Metrics Module and Introduce Data Processing Features
- Renamed DataProcessingModule for consistency.
- Removed unused MetricsController and MetricsService, along with their associated DTOs and tests.
- Introduced StateFetcherModule and StateFetcherService for real-time blockchain metrics polling.
- Added DataProcessingService to handle event processing and contract state management.
- Updated app.module.ts to reflect new module imports.
