+++ src/app.module.ts
@@ -14,7 +14,7 @@ import { TasksModule } from './tasks/tasks.module';
 import { ContractsModule } from './contracts/contracts.module';
 import { BlockchainsModule } from './blockchains/blockchains.module';
 import { MetricsModule } from './metrics/metrics.module';
-import { DataProcesingModule } from './data-procesing/data-procesing.module';
+import { DataProcesingModule } from './data-processing/data-processing.module';
 import { EventFetcherModule } from './event-fetcher/event-fetcher.module';
 import { StateFetcherModule } from './state-fetcher/state-fetcher.module';
