+++ src/data-processing/data-processing.module.ts
@@ -1,0 +1,4 @@
+import { Module } from '@nestjs/common';
+
+@Module({})
+export class DataProcesingModule {}
