+++ src/data-processing/data-processing.service.spec.ts
@@ -1,0 +1,18 @@
+import { Test, TestingModule } from '@nestjs/testing';
+import { DataProcessingService } from './data-processing.service';
+
+describe('EventFetcherService', () => {
+  let service: DataProcessingService;
+
+  beforeEach(async () => {
+    const module: TestingModule = await Test.createTestingModule({
+      providers: [DataProcessingService],
+    }).compile();
+
+    service = module.get<DataProcessingService>(DataProcessingService);
+  });
+
+  it('should be defined', () => {
+    expect(service).toBeDefined();
+  });
+});
