--- src/metrics/dto/create-metric.dto.ts
@@ -1,1 +1,0 @@
-export class CreateMetricDto {}
