--- src/metrics/dto/update-metric.dto.ts
@@ -1,4 +1,0 @@
-import { PartialType } from '@nestjs/mapped-types';
-import { CreateMetricDto } from './create-metric.dto';
-
-export class UpdateMetricDto extends PartialType(CreateMetricDto) {}
