--- src/metrics/entities/metric.entity.ts
@@ -1,1 +1,0 @@
-export class Metric {}
