--- src/metrics/metrics.controller.spec.ts
@@ -1,20 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { MetricsController } from './metrics.controller';
-import { MetricsService } from './metrics.service';
-
-describe('MetricsController', () => {
-  let controller: MetricsController;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [MetricsController],
-      providers: [MetricsService],
-    }).compile();
-
-    controller = module.get<MetricsController>(MetricsController);
-  });
-
-  it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
