+++ src/metrics/metrics.service.ts
@@ -1,26 +1,4 @@
 import { Injectable } from '@nestjs/common';
-import { CreateMetricDto } from './dto/create-metric.dto';
-import { UpdateMetricDto } from './dto/update-metric.dto';
 
 @Injectable()
-export class MetricsService {
-  create(createMetricDto: CreateMetricDto) {
-    return 'This action adds a new metric';
-  }
-
-  findAll() {
-    return `This action returns all metrics`;
-  }
-
-  findOne(id: number) {
-    return `This action returns a #${id} metric`;
-  }
-
-  update(id: number, updateMetricDto: UpdateMetricDto) {
-    return `This action updates a #${id} metric`;
-  }
-
-  remove(id: number) {
-    return `This action removes a #${id} metric`;
-  }
-}
+export class MetricsService {}
